+++ sdk/storage/azqueue/queue/client.go
@@ -1,4 +1,4 @@
-package service
+package queue
 
 import (
 	"context"
