+++ sdk/storage/azqueue/queue/responses.go
@@ -4,7 +4,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-package service
+package queue
 
 import (
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azqueue/internal/generated"
